// app/generated/api/schemas/media.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { MediaType } from './mediaType';

export interface Media {
  /** X 上でメディアを一意に識別できるキー */
  mediaKey: string;
  /** メディアの種類 */
  type: MediaType;
  /**
   * メディアの URL
   * @minLength 1
   * @maxLength 2083
   */
  url: string;
  /**
   * メディアの幅
   * @minimum 0
   */
  width: number;
  /**
   * メディアの高さ
   * @minimum 0
   */
  height: number;
}